Move per-backend I/O logic from functions.py into IbisBackend
The free functions in chronify/ibis/functions.py branched on
backend.name to decide how to normalize timestamps and whether to
convert Arrow to pandas. That feature envy blocked per-backend fast
paths (e.g., Arrow-native ingest, Spark-native execution) because the
dispatcher owned the format choice.

Push read_table, read_query, and write_table onto IbisBackend as
concrete methods with _post_read_normalize and _prepare_write_data
hooks that SQLiteBackend and SparkBackend override. Delete functions.py
and update all callers to invoke backend methods directly.
