+++ src/chronify/ibis/sqlite_backend.py
@@ -3,16 +3,26 @@
 import sqlite3
 from datetime import datetime
 from pathlib import Path
-from typing import Any, cast
+from typing import Any, Sequence, cast
 
 import ibis
 import ibis.expr.types as ir
 import pandas as pd
 import pyarrow as pa
 from loguru import logger
+from pandas import DatetimeTZDtype
 
 from chronify.exceptions import ConflictingInputsError, InvalidOperation, InvalidParameter
-from chronify.ibis.base import IbisBackend, ObjectType, _is_ddl
+from chronify.ibis.base import (
+    DatetimeRanges,
+    IbisBackend,
+    ObjectType,
+    TimeBaseModel,
+    TimeDataType,
+    _DATETIME_RANGES,
+    _is_ddl,
+    _normalize_timestamps,
+)
 
 
 def _adapt_value(v: Any) -> Any:
@@ -219,6 +229,62 @@ class SQLiteBackend(IbisBackend):
     def _commit_if_needed(self) -> None:
         if not self._in_transaction:
             self._connection.con.commit()
+
+    def _post_read_normalize(self, df: pd.DataFrame, config: DatetimeRanges) -> None:
+        """SQLite stores timestamps as text; coerce back to datetime dtypes."""
+        _convert_database_output_for_datetime(df, config)
+
+    def _prepare_write_data(
+        self,
+        data: pd.DataFrame | pa.Table,
+        configs: Sequence[TimeBaseModel],
+    ) -> pd.DataFrame:
+        """SQLite cannot ingest Arrow directly; convert to pandas and coerce TZ→UTC text."""
+        if isinstance(data, pa.Table):
+            data = data.to_pandas()
+        data = _normalize_timestamps(data, configs)
+        copied = False
+        for config in configs:
+            if isinstance(config, _DATETIME_RANGES):
+                data, copied = _convert_database_input_for_datetime(data, config, copied)
+        return data
+
+
+def _convert_database_output_for_datetime(df: pd.DataFrame, config: DatetimeRanges) -> None:
+    """Convert DataFrame datetime columns after SQLite output."""
+    if config.time_column not in df.columns:
+        return
+
+    col = df[config.time_column]
+    if config.dtype == TimeDataType.TIMESTAMP_TZ:
+        if col.dtype == object:
+            df[config.time_column] = pd.to_datetime(col, utc=True)
+        elif isinstance(col.dtype, DatetimeTZDtype):
+            df[config.time_column] = col.dt.tz_convert("UTC")
+        else:
+            df[config.time_column] = col.dt.tz_localize("UTC")
+    else:
+        if col.dtype == object:
+            df[config.time_column] = pd.to_datetime(col, utc=False)
+
+
+def _convert_database_input_for_datetime(
+    df: pd.DataFrame, config: DatetimeRanges, copied: bool
+) -> tuple[pd.DataFrame, bool]:
+    """Convert DataFrame datetime columns for SQLite input (store as UTC)."""
+    if config.dtype == TimeDataType.TIMESTAMP_NTZ:
+        return df, copied
+
+    if not copied:
+        df = df.copy()
+        copied = True
+
+    if isinstance(df[config.time_column].dtype, DatetimeTZDtype):
+        df[config.time_column] = df[config.time_column].dt.tz_convert("UTC")
+    else:
+        df[config.time_column] = df[config.time_column].dt.tz_localize("UTC")
+
+    return df, copied
 
 
 def _infer_sqlite_path(connection: ibis.BaseBackend) -> str | None:
