+++ src/chronify/time_zone_localizer.py
@@ -8,7 +8,6 @@ import pandas as pd
 from pandas import DatetimeTZDtype
 
 from chronify.ibis.base import IbisBackend
-from chronify.ibis.functions import read_query
 from chronify.models import TableSchema, MappingTableSchema
 from chronify.time_configs import (
     DatetimeRangeBase,
@@ -343,7 +342,7 @@ class TimeZoneLocalizerByColumn(TimeZoneLocalizerBase):
             .distinct()
             .filter(table[self.time_zone_column].notnull())
         )
-        time_zones = read_query(self._backend, expr, self._from_schema.time_config)[
+        time_zones = self._backend.read_query(expr, self._from_schema.time_config)[
             self.time_zone_column
         ].to_list()
